Create Forgotten Language
A forgotten language is a language no longer in use. Fortunately some of the words of those languages are still in use in some of the phrases of currently used languages. 
We are given few words from those languages and some of the phrases from the modern languages.
we are to find if the words exist in any of the phrases. 
T- number of text case
N- number of words
K- number of phrases
sample input:
2
3 2
piygu ezyfo rzotm
1 piygu
6 tefwz tefwz piygu ezyfo tefwz piygu
4 1
kssdy tjzhy ljzym kegqz
4 kegqz kegqz kegqz vxvyj

output:
YES YES NO
NO NO NO YES