Update just next.cpp
DevG likes too much fun to do with numbers. Once his friend Arya came and gave him a challenge, he gave DevG an array of digits which is forming a number currently (will be called as given number). DevG was challanged to find the just next greater number which can be formed using digits of given number. Now DevG needs your help to find that just next greater number and win the challenge.
Input:
2
5
1 5 4 8 3
10
1 4 7 4 5 8 4 1 2 6

Output:
15834
1474584162


The idea behind the solution is to check if the array is already not in decreasing order
print "-1" if it is
otherwise starting from the right side(i.e.) end of the array look for the breakpoint 
i.e. element from where the array starts decreasing again
save this element in some variable(say temp) for now it will be needed later
now siince we have all the elements in are stack push it into an array  and sort it to find position of element just greater than temp
swap that element with the first element (element at 0th position)
sort the remaining array and push it back into the main array
print the main array



// time complexity - o(n)
//space complexity - o(n)       -not sure