Create maxRecAreaHisto.cpp
The question is to find the rectangle with maximum area in a given histogram. Consider following histogram{2, 1, 4, 5, 1, 3, 3,}

                                         ____
5-                               ____| ___|
4-                              |      |      |
3-                              |      |      |         ____ ____
2-                 ____       |      |      |       |       |      |
1-                |      |____|      |      | ____|       |      |
0- __________|____|____|____|___ |____ |_____|___ |________________________________________________
          
here max area is 4*2=8
Calculate the area of the largest rectangle in a histogram that is aligned at the common base line, too. The figure on the right shows the largest aligned rectangle for the depicted histogram.



the idea is to keep an increasing stack i.e, top element of the stack must contain a value greater than or equal to the element below it
thus the element just below the top will be the left boundary and i'th element in the histogram array will be the right limit if it is less than stack top